% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{AnalyzeResult}
\alias{AnalyzeResult}
\title{Analysis the results of CCNMF}
\usage{
AnalyzeResult(
  RNAmatrix,
  CNVmatrix,
  ncluster = 3,
  lambda1,
  lambda2,
  QC = TRUE,
  GeneFilter = TRUE,
  reference_name = "hg38"
)
}
\description{
Analysis the results of CCNMF
}
