% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_pre_processing.R
\name{Center_cells_across_chromosome}
\alias{Center_cells_across_chromosome}
\title{Centering cells after smoothing}
\usage{
Center_cells_across_chromosome(data, method = "median")
}
\arguments{
\item{data}{the input matrix}

\item{method}{using 'median' or 'mean' for centering}
}
\value{
matrix after centering among cells across chromosomes
}
\description{
Centering cells after smoothing
}
