% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDNA_pre_processing.R
\name{Convert_scDNA_genes_matrix}
\alias{Convert_scDNA_genes_matrix}
\title{Convert CNV on gene level}
\usage{
Convert_scDNA_genes_matrix(scdna_matrix, gene_locs)
}
\arguments{
\item{scdna_matrix}{The bins times cells CNV matrix}

\item{gene_locs}{the list of bins index on genes}
}
\value{
the CNV matrix on genes
}
\description{
Merge the bins corresponding to the same gene
}
