% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultpar_CCNMF.R
\name{default_parameters_hyper}
\alias{default_parameters_hyper}
\title{Default hyper parameters}
\usage{
default_parameters_hyper(
  CNVmatrix,
  W10,
  H10,
  RNAmatrix,
  W20,
  H20,
  CoupledMatrix,
  beta,
  alpha
)
}
\arguments{
\item{CNVmatrix}{the copy number matrix}

\item{W10}{W matrix of CNV's NMF}

\item{H10}{H matrix of CNV's NMF}

\item{RNAmatrix}{gene expression matrix}

\item{W20}{W matrix of RNA NMF}

\item{H20}{H matrix of RNA's NMF}

\item{CoupledMatrix}{coupled Matrix}

\item{beta}{hyper-parmater}

\item{alpha}{hyper-parmater}
}
\description{
Initialize the parameters lambda1 and lambda2 by hyper-parameter alpha and beta
}
