% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultpar_CCNMF.R
\name{default_parameters_order}
\alias{default_parameters_order}
\title{Default parameters by order}
\usage{
default_parameters_order(
  CNVmatrix,
  W10,
  H10,
  RNAmatrix,
  W20,
  H20,
  CoupledMatrix
)
}
\arguments{
\item{CNVmatrix}{the copy number matrix}

\item{W10}{W matrix of CNV's NMF}

\item{H10}{H matrix of CNV's NMF}

\item{RNAmatrix}{gene expression matrix}

\item{W20}{W matrix of RNA NMF}

\item{H20}{H matrix of RNA's NMF}

\item{CoupledMatrix}{coupled Matrix}
}
\description{
Initialize the parameters lambda1 and lambda2 by seting the items in object function in the same order
}
