% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccNMF.R
\name{eigenMapMatMult}
\alias{eigenMapMatMult}
\title{Define multiplication between matrices}
\usage{
eigenMapMatMult(A, B)
}
\arguments{
\item{A}{Matrix}

\item{B}{Matrix}
}
\description{
Define multiplication between matrices
}
