% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{Estimate_Coupled_matrix}
\alias{Estimate_Coupled_matrix}
\title{Converting region-based copy number to gene based
Find the corresponding cnv regions and gene bins according to the reference genome.
This function refers as below:
https://kieranrcampbell.github.io/clonealign/preparing_copy_number_data.html}
\usage{
Estimate_Coupled_matrix(RNAmatrix, CNVmatrix, reference_name = "hg19")
}
\arguments{
\item{RNAmatrix}{the gene expression matrix which is a matrix file}

\item{CNVmatrix}{the copy number varients matrix which is a dataframe file}

\item{reference_name}{the name of reference annotations, default is 'hg19'}
}
\value{
RNAmatrix_match in which the number of genes equals to the number of cnv regions in
 CNVmatrix_match. More details, the genes in RNAmatrix_match is one-to-one correspondence with the
 cnv regions in CNVmatrix_match.
}
\description{
Converting region-based copy number to gene based
Find the corresponding cnv regions and gene bins according to the reference genome.
This function refers as below:
https://kieranrcampbell.github.io/clonealign/preparing_copy_number_data.html
}
