% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_pre_processing.R
\name{Filter_genes}
\alias{Filter_genes}
\title{Filter genes with less express cells}
\usage{
Filter_genes(mat, min_num_cells)
}
\arguments{
\item{mat}{the scRNA-seq matrix}

\item{min_num_cells}{the minmum number of cells}
}
\value{
matrix after filtering genes
}
\description{
Filter genes with less express cells
}
