% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_pre_processing.R
\name{Filter_genes_below_mean_exp_cutoff}
\alias{Filter_genes_below_mean_exp_cutoff}
\title{Filter genes that expressed fewer than certain number of cells}
\usage{
Filter_genes_below_mean_exp_cutoff(mat, cutoff)
}
\arguments{
\item{mat}{the input scRNA matrix}

\item{cutoff}{the percentage of filtering genes}
}
\value{
matrix after filtering genes
}
\description{
Filter genes that expressed fewer than certain number of cells
}
