% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDNA_pre_processing.R
\name{Iden_replicating_cells}
\alias{Iden_replicating_cells}
\title{Identify replicating cells}
\usage{
Iden_replicating_cells(scdna_matrix_merge)
}
\arguments{
\item{scdna_matrix_merge}{1Mb-segments copy number matrix}
}
\value{
the index of non-replicating cells
}
\description{
Identify the replicating/noisy cells by computing standard deviation for cells on 50kb segemental-bins
}
