% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDNA_pre_processing.R
\name{Iden_signal_segments}
\alias{Iden_signal_segments}
\title{CNV feature selection}
\usage{
Iden_signal_segments(scdna_matrix)
}
\arguments{
\item{scdna_matrix}{the 50kb-segmental bins * non-replicating cells matrix}
}
\value{
The index of selecting signal segments
}
\description{
Identify the signal segments by computing standard deviation for 50kb bins across cells
}
