% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input.R
\name{InputDNA}
\alias{InputDNA}
\title{Input the copy number copy in scDNA-seq data, which is a csv file.}
\usage{
InputDNA(pathDNA, file_CNV, Verbose = TRUE)
}
\arguments{
\item{pathDNA}{the path of the CNV file located}

\item{file_CNV}{the name of the CNV file. csv file is suggested.}

\item{Verbose}{the CNV files is cnvcalls of cellrangers: Otherwise, the Verbose is FALSE when the cnv file is .csv file as the format of example data.}
}
\value{
The CNV matrix in which the first four rows are "chr", "start", "end", "width". The rest of the rows are
the single cell samples.
}
\description{
Input the copy number copy in scDNA-seq data, which is a csv file.
}
