% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{InputRNA}
\alias{InputRNA}
\title{Input the gene expression matrix which is from 10X}
\usage{
InputRNA(
  pathRNA,
  file_gene = "genes.tsv",
  file_barcodes = "barcodes.tsv",
  file_matrix = "matrix.mtx"
)
}
\arguments{
\item{pathRNA}{the path of the standard three files: barcodes.tsv, genes.tsv, matrix.mtx located}

\item{file_gene}{the name of the gene file. Default is 'genes.tsv'}

\item{file_barcodes}{the name of the barcode file. Default is 'barcodes.tsv'}

\item{file_matrix}{the name of the gene counts matrix. Default is 'matrix.mtx'}
}
\value{
The inputed gene expression matrix in which rows are cells, columns are genes.
}
\description{
Input the gene expression matrix which is from 10X
}
