% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overall_CCNMF.R
\name{Integrate_CCNMF}
\alias{Integrate_CCNMF}
\title{Integrate CCNMF}
\usage{
Integrate_CCNMF(
  CNVmatrix_input,
  RNAmatrix_input,
  ncluster,
  CoupledMatrix = NULL,
  initial_parameters = c("hyper-parameter", "same-order", "user-defined"),
  lambda1 = 1,
  lambda2 = 2,
  alpha = NULL,
  beta = NULL,
  initial_coupling = c("default", "user-defined")
)
}
\arguments{
\item{CNVmatrix_input}{input cnv matrix}

\item{RNAmatrix_input}{inout rna matrix}

\item{ncluster}{the number of given clusters}

\item{CoupledMatrix}{the user-defined coupled matrix}

\item{initial_parameters}{the set of lambda1 and lambda2}

\item{lambda1}{the weight of NMF}

\item{lambda2}{the weight of coupled item}

\item{alpha}{hyper-parameter alpha}

\item{beta}{hyper-parameter beta}

\item{initial_coupling}{set up coupling matrix}
}
\value{
list of results
}
\description{
whole progress of running CCNMF
}
