% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_pre_processing.R
\name{Invert_logtransform}
\alias{Invert_logtransform}
\title{Invert Log transformation}
\usage{
Invert_logtransform(data)
}
\arguments{
\item{data}{input matrix}
}
\value{
the matrix after invert log transformation
}
\description{
Invert Log transformation
}
