% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_pre_processing.R
\name{Logtransform}
\alias{Logtransform}
\title{Log transformation}
\usage{
Logtransform(data)
}
\arguments{
\item{data}{the input matrix}
}
\value{
the log2 normalized matrix
}
\description{
Log transformation
}
