% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_pre_processing.R
\name{Max_threshold_bounds}
\alias{Max_threshold_bounds}
\title{Apply max/min threshold bounds}
\usage{
Max_threshold_bounds(data, threshold)
}
\arguments{
\item{data}{input matrix}

\item{threshold}{Set the max and minmum bounds for matrix}
}
\value{
bounded matrix
}
\description{
Apply max/min threshold bounds
}
