% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDNA_pre_processing.R
\name{Merge_bins2segments}
\alias{Merge_bins2segments}
\title{Merge CNV bins into genome segment}
\usage{
Merge_bins2segments(scdna_matrix, bin_size = 50)
}
\arguments{
\item{scdna_matrix}{name of copy number variation file}

\item{bin_size}{number of #kb bins}
}
\value{
scdna_matrix_merge  CNV matrix with 1Mb segments
}
\description{
Merge small bins to large segement
}
