% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccNMF.R
\name{neighbor_default}
\alias{neighbor_default}
\title{Dividing nodes in the network into communities makes connections within communities tight and connections between communities sparse
The algorithm keeps merging nodes with similar connectivity patterns in an iterative manner until it reaches the optimal community segmentation result}
\usage{
neighbor_default(X)
}
\arguments{
\item{X}{Matrix}
}
\description{
Dividing nodes in the network into communities makes connections within communities tight and connections between communities sparse
The algorithm keeps merging nodes with similar connectivity patterns in an iterative manner until it reaches the optimal community segmentation result
}
