% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_pre_processing.R
\name{normalize_counts_by_seq_depth}
\alias{normalize_counts_by_seq_depth}
\title{Normalizes count data by total sum scaling}
\usage{
normalize_counts_by_seq_depth(mat, normalize_factor = NA)
}
\arguments{
\item{mat}{the input scRNA matrix}

\item{normalize_factor}{the scale factor for normalization with depth}
}
\value{
matrix after normaliztion by scale factor
}
\description{
Normalizes count data by total sum scaling
}
