% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_pre_processing.R
\name{Normalize_RNA}
\alias{Normalize_RNA}
\title{Normalization for single-cell RNA-seq data}
\usage{
Normalize_RNA(scRNA_raw_matrix, gene_locs)
}
\arguments{
\item{scRNA_raw_matrix}{the scRNA-seq matrix with raw counts and WGS genes}

\item{gene_locs}{the genome reference with gene, chr, start, end}
}
\value{
normalized scRNA matrix
}
\description{
Normalization for single-cell RNA-seq data
}
