% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plot}
\alias{Plot}
\title{The function of plotting figure}
\usage{
Plot(Data, Clones, title, labelx, labely)
}
\arguments{
\item{Data}{the input data which need to be plotted, which can be raw matrix or H matrix concluded by NMF}

\item{Clones}{the clusters label}

\item{title}{the title of the figure, which is a string}

\item{labelx}{X-coordinate of name, such as 'pca 1', 'tsne 1', 'umap 1'}

\item{labely}{Y-coordinate of name, such as 'pca 2', 'tsne 2', 'umap 2'}
}
\value{
A pdf file
}
\description{
The function of plotting figure
}
