% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_CNV_WGS_heatmap.R
\name{Plot_CNV_heatmap}
\alias{Plot_CNV_heatmap}
\title{Plot CNV Heatmap}
\usage{
Plot_CNV_heatmap(mat, labels, gene_chr_mark, filename, is.noise)
}
\arguments{
\item{mat}{matrix of copy number}

\item{labels}{subclones clustering labels}

\item{gene_chr_mark}{chromosome regions}

\item{filename}{the name of output figure}

\item{is.noise}{Logistic parm. if TRUE, the subclone figure contains noise cells}
}
\value{
the pdf figure
}
\description{
Plot CNV Heatmap
}
