% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_CNV_WGS_heatmap.R
\name{Plot_CNV_subclone_heatmap}
\alias{Plot_CNV_subclone_heatmap}
\title{Plot Subclonal CNV heatmap}
\usage{
Plot_CNV_subclone_heatmap(
  CNVsample,
  label,
  Gene_Order,
  index_remain_cells = NULL,
  is.noise = FALSE
)
}
\arguments{
\item{CNVsample}{copy number data matrix with all genes, dimension: cells}

\item{label}{subclones labeling of CCNMF}

\item{Gene_Order}{dataframe of genes locs}

\item{index_remain_cells}{cell index of non-replicating cells}

\item{is.noise}{Logistic parm. if TRUE, the subclone figure contains noise cells}
}
\value{
the figure of subclones heatmap
}
\description{
Plot Subclonal CNV heatmap
}
