% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plot_heatmap}
\alias{Plot_heatmap}
\title{Plot the heatmap for differential genes in scRNA-seq data}
\usage{
Plot_heatmap(
  Data,
  label,
  P,
  Datatype = "DNA",
  title = "The heatmap of differential expression in scRNA-seq data"
)
}
\arguments{
\item{Data}{scRNA-seq gene expression matrix}

\item{label}{the clustering label of scRNA-seq data}

\item{P}{the p_values matrix by t test the DE genes among clusters in scRNA-seq data}

\item{title}{a string, default as 'The heatmap of differential expression in scRNA-seq data}
}
\value{
D_assign: a matrix that rows are the clusters and cloumns are the DE genes of all clusters

A pdf file which is the heatmap figure
}
\description{
Plot the heatmap for differential genes in scRNA-seq data
}
