% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotMainResult}
\alias{PlotMainResult}
\title{Output the all visualization of results from CCNMF
Especially, the paired heatmap of differential genes and dimension reduction for both scRNA-seq and scDNA-seq data}
\usage{
PlotMainResult(CNVmatrix, RNAmatrix, Result_CCNMF, DElist)
}
\arguments{
\item{CNVmatrix}{copy number matrix}

\item{RNAmatrix}{gene expression matrix}

\item{DElist}{The list of differentiate genes}

\item{Results_CCNMF}{The result of CCNMF}
}
\value{
The integrated figure
}
\description{
Output the all visualization of results from CCNMF
}
