% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plotpca}
\alias{Plotpca}
\title{Plot the dimensional figure for scRNA-seq and scDNA-seq datasets
The rows in Data matrix represent the cells, the columns represent the genes/chr bins}
\usage{
Plotpca(Data, label, title, Datatype = "scRNA-seq")
}
\arguments{
\item{Data}{the orgial matrix or H matrix when the raw matrix after NMF}

\item{label}{the clusters label of this figure which is the result of CCNMF}

\item{title}{the title of the figure}

\item{Datatype}{The type of input Data matrix, 'scRNA-seq' or 'scDNA-seq'}
}
\value{
The pdf figure based on PCA
}
\description{
Plot the dimensional reduction figure by PCA
}
