% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plottsne}
\alias{Plottsne}
\title{Plot the dimensional reduction figure by tsne
Before tsne, the PCA is applied to the original Data, then select the top 15 PCs and apply
tsne to these components.}
\usage{
Plottsne(Data, label, title, Datatype = "scRNA-seq", need_PCA = TRUE)
}
\arguments{
\item{Data}{the orgial matrix or H matrix when the raw matrix after NMF}

\item{label}{the clusters label of this figure which is the result of CCNMF}

\item{Datatype}{The type of input Data matrix, 'scRNA-seq' or 'scDNA-seq'}
}
\value{
The pdf figure based on tsne
}
\description{
Plot the dimensional reduction figure by tsne
Before tsne, the PCA is applied to the original Data, then select the top 15 PCs and apply
tsne to these components.
}
