% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plotumap}
\alias{Plotumap}
\title{Plot the dimensional reduction figure by umap
Firstly, we apply PCA for the high-dimensional data, then use Umap to the top 15 PCs.}
\usage{
Plotumap(Data, label, Datatype = "scRNA-seq", need_PCA = TRUE)
}
\arguments{
\item{Data}{the orgial matrix or H matrix when the raw matrix after NMF}

\item{label}{the clusters label of this figure which is the result of CCNMF}

\item{Datatype}{The type of input Data matrix, 'scRNA-seq' or 'scDNA-seq'}

\item{need_PCA}{Logic}
}
\value{
The pdf figure based on umap
}
\description{
Plot the dimensional reduction figure by umap
Firstly, we apply PCA for the high-dimensional data, then use Umap to the top 15 PCs.
}
