% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{process_RNA_matrix}
\alias{process_RNA_matrix}
\title{Preprocessed scRNA matrix by Seurat including Normalization, scaling and selcting high variable genes.
Input the path where the 10X RNA-seq located}
\usage{
process_RNA_matrix(pathRNA)
}
\description{
Preprocessed scRNA matrix by Seurat including Normalization, scaling and selcting high variable genes.
Input the path where the 10X RNA-seq located
}
