% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{QualityControl}
\alias{QualityControl}
\title{Quality control the cells in scRNA-seq matrix}
\usage{
QualityControl(RNAmatrix)
}
\arguments{
\item{The}{gene expression matrix of scRNA-seq}
}
\value{
The number of cells will be less
}
\description{
Quality control the cells in scRNA-seq matrix
}
