% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccNMF.R
\name{readRawData}
\alias{readRawData}
\title{Read raw data in 10X format
X1: scRNA-seq data
X2: scATAC-seq data
PeakName: Name of the ATAC peaks
Barcode: Cell Name
GeneLoc: Location of genes on chromosomes
GeneName: Gene name
PeakLoc: Location of peaks on chromosomes
chr: Chromosomes name}
\usage{
readRawData(foldername)
}
\arguments{
\item{foldername}{Data storage address, required to contain features.tsv, barcodes.tsv and matrix.mtx}
}
\description{
Read raw data in 10X format
X1: scRNA-seq data
X2: scATAC-seq data
PeakName: Name of the ATAC peaks
Barcode: Cell Name
GeneLoc: Location of genes on chromosomes
GeneName: Gene name
PeakLoc: Location of peaks on chromosomes
chr: Chromosomes name
}
