% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCNMF.R
\name{run_CCNMF}
\alias{run_CCNMF}
\title{Run CCNMF function
Run \code{CCNMF} accross a range pf initializations}
\usage{
run_CCNMF(ncluster, RNAmatrix, CNVmatrix, CoupledMatrix, lambda1, lambda2)

run_CCNMF(ncluster, RNAmatrix, CNVmatrix, CoupledMatrix, lambda1, lambda2)
}
\arguments{
\item{ncluster}{the number of subpopulations/clones seted by users}

\item{RNAmatrix}{A matrix of gene expression in scRNA-seq}

\item{CNVmatrix}{A matrix of copy number in scDNA-seq}

\item{CoupledMatrix}{A coupled matrix between genes and cnv bins}

\item{lambda1}{A hyper-parameter which is seted by users}

\item{lambda2}{A hyper-parameter which is seted by users}
}
\value{
The list(H1, H2, S1, S2)

The list(H1, H2, S1, S2)
}
\description{
Run CCNMF function
Run \code{CCNMF} accross a range pf initializations

Run CCNMF function
Run \code{CCNMF} accross a range pf initializations
}
\details{
This function is core of CCNMF and the parameters lambda1 and lambda2 control the objective function, meanwhile, the two hyper-parameters can be
initially assigend.

This function is core of CCNMF and the parameters lambda1 and lambda2 control the objective function, meanwhile, the two hyper-parameters can be
initially assigend.
}
