% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{run_Seurat_RNA}
\alias{run_Seurat_RNA}
\title{Preprocessing single-cell RNA-seq data by Seurat including normalization, scale, select high
variable features.}
\usage{
run_Seurat_RNA(RNAdata, min.cells = 6, min.features = 0)
}
\description{
Preprocessing single-cell RNA-seq data by Seurat including normalization, scale, select high
variable features.
}
