% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccNMF.R
\name{Simbased_clustering}
\alias{Simbased_clustering}
\title{Realizing cell clustering}
\usage{
Simbased_clustering(H, k = 20)
}
\arguments{
\item{H}{Matrix for integrating results}

\item{k}{The number of nearest neighbors chosen by the KNN}
}
\description{
Realizing cell clustering
}
