% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_pre_processing.R
\name{Smooth_by_chromosome}
\alias{Smooth_by_chromosome}
\title{Smooth the matrix on chromosomes}
\usage{
Smooth_by_chromosome(data, window_length, gene_locs, smooth_end = TRUE)
}
\arguments{
\item{data}{the input matrix}

\item{window_length}{the number of genes}

\item{gene_locs}{the geneom reference}

\item{smooth_end}{if set the smooth end or not}
}
\value{
matrix after smooth on chromosomes
}
\description{
Smooth the matrix on chromosomes
}
