% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_pre_processing.R
\name{smooth_helper}
\alias{smooth_helper}
\title{Smoothing by chromosome through moving average}
\usage{
smooth_helper(obs_data, window_length)
}
\arguments{
\item{obs_data}{the input matrix}

\item{window_length}{the number of genes as window for smoothing}
}
\value{
matrix after smoothing
}
\description{
Smoothing by chromosome through moving average
}
