% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_pre_processing.R
\name{smooth_window}
\alias{smooth_window}
\title{make the window smooth for matrix}
\usage{
smooth_window(data, window_length)
}
\arguments{
\item{data}{the input matrix}

\item{window_length}{the number of genes for window smooth}
}
\value{
matrix after smooth normaliztion
}
\description{
make the window smooth for matrix
}
