% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_CNV_WGS_heatmap.R
\name{Sort_subclones}
\alias{Sort_subclones}
\title{Sort subclones}
\usage{
Sort_subclones(mat, labels, hc)
}
\arguments{
\item{mat}{the matrix of segments-bins}

\item{labels}{subclones index for cells}

\item{hc}{results of hierarchical clustering}
}
\value{
list of ordered matrix and cluster label
}
\description{
Sort subclones
}
