% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_CNV_WGS_heatmap.R
\name{SortCells}
\alias{SortCells}
\title{Sort cells}
\usage{
SortCells(tmp_scdna_matrix_arm, S1, Corr)
}
\arguments{
\item{tmp_scdna_matrix_arm}{scDNA matrix}

\item{S1}{labels of subclonses}

\item{Corr}{the distance to normal cell}
}
\value{
Ordered cells based on per subclones
}
\description{
Sort cells
}
