% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_CNV_WGS_heatmap.R
\name{SortCells_in_subclone}
\alias{SortCells_in_subclone}
\title{Sort cells within subclone}
\usage{
SortCells_in_subclone(tmp_scdna_matrix, Corr)
}
\arguments{
\item{tmp_scdna_matrix}{the matrix of cells}

\item{Corr}{the distance to normal cell}
}
\value{
Ordered cells matrix
}
\description{
Sort cells within subclone
}
