% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_pre_processing.R
\name{Subtract_ref_mean_exp}
\alias{Subtract_ref_mean_exp}
\title{Subtract average reference}
\usage{
Subtract_ref_mean_exp(data, inv_log = FALSE)
}
\arguments{
\item{data}{the input matrix}

\item{inv_log}{if perform invert log transformation or not}
}
\value{
the matrix after subtract the average expression
}
\description{
Subtract average reference
}
